Implemented adjacency resolution. We maintain a data structure (ResolutionMap) for that purpose throughout all stages of the algorithm. It maps each move to a list of adjacencies it resolves. We update the map whenever: - a move is removed from refs as a result of an incompatible word being placed - a new adjacency is created (by placing a word onto the grid)
Whenever the resolution map is not empty, we only make resolving moves.

Note that in this version we don't fail when there is no resolver for an
adjacency. This clearly works better because the "AI" is able to learn
from moves even if they do not result in a valid state (in conjunction
with other moves).